api/requirements.txt
decorator==5.1.1
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
psycopg2==2.9.3
PyJWT==2.5.0
requests==2.28.1
SQLAlchemy==1.4.41
validators==0.20.0
watchdog==2.1.9
Werkzeug==2.2.2
gunicorn==20.1.0
